chore(Lytics): move sdk loader to different file
